+++ include/swift/DependencyScan/DependencyScanningTool.h
@@ -87,14 +87,15 @@ public:
   /// occurred, \c swiftscan_dependency_result_t otherwise.
   llvm::ErrorOr<swiftscan_dependency_graph_t>
   getDependencies(ArrayRef<const char *> Command,
-                  const llvm::StringSet<> &PlaceholderModules);
+                  const llvm::StringSet<> &PlaceholderModules,
+                  StringRef WorkingDirectory);
 
   /// Collect the set of imports for the input module
   ///
   /// \returns a \c StringError with the diagnostic output if errors
   /// occurred, \c swiftscan_prescan_result_t otherwise.
   llvm::ErrorOr<swiftscan_import_set_t>
-  getImports(ArrayRef<const char *> Command);
+  getImports(ArrayRef<const char *> Command, StringRef WorkingDirectory);
 
   /// Collect the full module dependency graph for the input collection of
   /// module names (batch inputs) and output them to the
@@ -104,7 +105,8 @@ public:
   std::vector<llvm::ErrorOr<swiftscan_dependency_graph_t>>
   getDependencies(ArrayRef<const char *> Command,
                   const std::vector<BatchScanInput> &BatchInput,
-                  const llvm::StringSet<> &PlaceholderModules);
+                  const llvm::StringSet<> &PlaceholderModules,
+                  StringRef WorkingDirectory);
 
   /// Writes the current `SharedCache` instance to a specified FileSystem path.
   void serializeCache(llvm::StringRef path);
@@ -121,13 +123,15 @@ public:
   /// Using the specified invocation command, instantiate a CompilerInstance
   /// that will be used for this scan.
   llvm::ErrorOr<ScanQueryInstance>
-  initCompilerInstanceForScan(ArrayRef<const char *> Command);
+  initCompilerInstanceForScan(ArrayRef<const char *> Command,
+                              StringRef WorkingDirectory);
 
 private:
   /// Using the specified invocation command, initialize the scanner instance
   /// for this scan. Returns the `CompilerInstance` that will be used.
   llvm::ErrorOr<ScanQueryInstance>
-  initScannerForAction(ArrayRef<const char *> Command);
+  initScannerForAction(ArrayRef<const char *> Command,
+                       StringRef WorkingDirectory);
 
   /// Shared cache of module dependencies, re-used by individual full-scan queries
   /// during the lifetime of this Tool.
